+++ CHANGELOG.md
@@ -41,6 +41,9 @@ Main (unreleased)
   - `prometheus.exporter.mysql` collects metrics from a MySQL database.
     (@spartan0x117)
   - `prometheus.exporter.postgres` collects metrics from a PostgreSQL database. (@spartan0x117)
+  - `otelcol.auth.sigv4` performs AWS Signature Version 4 (SigV4) authentication 
+    for making requests to AWS services via `otelcol` components that support
+    authentication extensions. (@ptodev)
 
 ### Enhancements